+++ components/home/Hero/Hero.tsx
@@ -1,7 +1,7 @@
-import { HeroProps } from "@/types/data";
-import Typography from "@/components//base/Typography/Typography";
-import Link from "@/components/base/Link/Link";
-import { HeroLinkIcon } from "@/utils/data.utils";
+import { HeroProps } from "../../../types/data";
+import Typography from "../../base/Typography/Typography";
+import Link from "../../base/Link/Link";
+import { HeroLinkIcon } from "../../../utils/data.utils";
 
 const Hero = ({ headline, links, summary }: HeroProps) => {
   return (
