+++ components/home/Projects/Projects.tsx
@@ -1,9 +1,9 @@
 "use client";
-import { ProjectCardProps } from "@/types/data";
-import Button from "@/components/base/Button/Button";
-import Typography from "@/components/base/Typography/Typography";
+import { ProjectCardProps } from "../../../types/data";
+import Button from "../../base/Button/Button";
+import Typography from "../../base/Typography/Typography";
 import ProjectCard from "./ProjectCard";
-import useExpand from "@/hooks/useExpand";
+import useExpand from "../../../hooks/useExpand";
 
 const Projects = ({
   title,
