+++ components/layout/Header/ThemeButton.tsx
@@ -1,7 +1,7 @@
 import { FaRegMoon } from "react-icons/fa";
 import { FiSun } from "react-icons/fi";
-import Button from "@/components/base/Button/Button";
-import { useTheme } from "@/hooks/useTheme";
+import Button from "../../base/Button/Button";
+import { useTheme } from "../../../hooks/useTheme";
 
 const ThemeButton = () => {
   const { theme, toggleTheme } = useTheme();
