+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sharonyang16.github.io",
+  "name": "sharonyyang.com",
   "version": "0.1.0",
   "private": true,
   "scripts": {
