+++ utils/data.utils.tsx
@@ -1,7 +1,7 @@
 import { FaFigma, FaGithub, FaLinkedin } from "react-icons/fa";
 import { FaCode } from "react-icons/fa6";
 import { MdArrowOutward, MdEmail } from "react-icons/md";
-import { LinkIconProps } from "@/types/data";
+import { LinkIconProps } from "../types/data";
 
 export const HeroLinkIcon = ({ name }: LinkIconProps) => {
   switch (name) {
